RavenDB-22200 - fixing PR comments#1
